+++ common/src/main/java/net/mehvahdjukaar/polytone/content/entity/ParticleSpawnRecord.java
@@ -1,7 +1,6 @@
 package net.mehvahdjukaar.polytone.content.entity;
 
-import net.minecraft.world.phys.Vec3;
 import org.joml.Matrix4fc;
 
-public record ParticleSpawnRecord(Matrix4fc transform, EntityParticleEmitter emitter) {
+public record ParticleSpawnRecord(Matrix4fc matrix, EntityParticleEmitter emitter) {
 }